wasm-bindgen-test-runner: Updated feature tests to use a more explicit step in terms of the content of the assembly.
