+++ tests/wasm_bindgen_test_runner/__features__/invocation/with_an_assembly/with_arguments/__skip_eq_pattern/count_2/level_1/with_two_partial_match_successful_tests/skips_tests_feature.rs
@@ -1,4 +1,4 @@
-use crate::__steps__::assembly::given_there_is_an_assembly_with_two_successful_level_1_tests;
+use crate::__steps__::assembly::given_there_is_an_assembly_with;
 use crate::__steps__::standard_error::then_the_standard_error_should_be_empty;
 use crate::__steps__::standard_output::then_the_standard_output_should_have;
 use crate::__steps__::standard_output::then_the_standard_output_should_not_have;
@@ -11,7 +11,23 @@ use auroka_morpheus_macros_feature::feature;
 feature! {
     test_mode: TestMode
 
-    given_there_is_an_assembly_with_two_successful_level_1_tests();
+    given_there_is_an_assembly_with(r#"
+mod level_1 {
+    use wasm_bindgen_test::*;
+
+    #[wasm_bindgen_test]
+    fn pass_1() {
+        console_log!("pass_1 standard output");
+        assert_eq!(1, 1);
+    }
+
+    #[wasm_bindgen_test]
+    fn pass_2() {
+        console_log!("pass_2 standard output");
+        assert_eq!(1, 1);
+    }
+}
+"#);
     when_wasm_bindgen_test_runner_is_invoked_with_the_assembly_for_test_mode_and_the_arguments(test_mode, "--skip=pass_1 --skip=level_1");
 
     "Outputs no information about the skipped test 1" {
