+++ tests/wasm_bindgen_test_runner/__steps__/assembly/mod.rs
@@ -3,11 +3,9 @@ mod given_there_is_an_assembly_with;
 mod given_there_is_an_assembly_without_anything;
 mod with_one_successful_and_one_failing_tests;
 mod with_one_successful_and_one_ignored_tests;
-mod with_two_successful_tests;
 
 pub use assembly_builder::*;
 pub use given_there_is_an_assembly_with::*;
 pub use given_there_is_an_assembly_without_anything::*;
 pub use with_one_successful_and_one_failing_tests::*;
 pub use with_one_successful_and_one_ignored_tests::*;
-pub use with_two_successful_tests::*;
